Set cookie after login/createShop  (#10)
* design the entities

* design the entities

* redesign the entities...

* Add Mapper

* edit

* some changes

* finish the mapper files

* part of service class complete

* ssss

* handle controller

* commit

* fix errors

* remove the useless files

* some revise

* shop and photo part

* wo ri

* revise the shop

* almost complete

* finish

* revise

* result

* add pom.xml

* add pom.xml

* bug fix

* owner finish

* owner finish

* message

* shop bugs fixed

* product finished

* all bugs fixed

* bug fix

* Bugs fix

* add new functions

* finish

* finish

* merge

* merge

* merge

* update module

* update module

* some rebuild

* merge

* Add avalidateCode

* add cookie

+++ src/main/java/com/heitian/ssm/controller/ProductController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -58,7 +59,10 @@ public class ProductController {
      */
     @ResponseBody
     @RequestMapping("/add")
-    public Result addProduct(@RequestBody ProductBo productBo) {
+    public Result addProduct(@RequestBody ProductBo productBo,
+                             @CookieValue(value = "OwnerEmail",defaultValue = "swc") String email) {
+        if("swc".equals(email))
+            return returnResult();
         return productService.addProduct(productBo);
     }
 
@@ -69,7 +73,10 @@ public class ProductController {
      */
     @ResponseBody
     @RequestMapping("/delete")
-    public Result deleteProduct(@RequestBody ProductBo productBo) {
+    public Result deleteProduct(@RequestBody ProductBo productBo,
+                                @CookieValue(value = "OwnerEmail",defaultValue = "swc") String email) {
+        if("swc".equals(email))
+            return returnResult();
         return productService.deleteProduct(productBo);
     }
 
@@ -80,7 +87,10 @@ public class ProductController {
      */
     @ResponseBody
     @RequestMapping("/update")
-    public Result updateProduct(@RequestBody ProductBo productBo) {
+    public Result updateProduct(@RequestBody ProductBo productBo,
+                                @CookieValue(value = "OwnerEmail",defaultValue = "swc") String email) {
+        if("swc".equals(email))
+            return returnResult();
         return productService.updateProduct(productBo);
     }
 
@@ -93,7 +103,10 @@ public class ProductController {
      */
     @ResponseBody
     @RequestMapping(value="/searchByOwner",method= RequestMethod.GET)
-    public List<ProductBo> searchByOwner(@RequestParam long ownerId, @RequestParam int page, @RequestParam int pageNum) {
+    public List<ProductBo> searchByOwner(@RequestParam long ownerId, @RequestParam int page, @RequestParam int pageNum,
+                                         @CookieValue(value = "OwnerEmail",defaultValue = "swc") String email) {
+        if("swc".equals(email))
+            return new ArrayList<>();
         return productService.searchProductBosByOwner(ownerId,page,pageNum);
     }
 
@@ -108,6 +121,13 @@ public class ProductController {
         Result result =new Result();
         result.setMessage(Integer.toString(productService.getOwnerProductCount(ownerId)));
         result.setStatus(1);
+        return result;
+    }
+
+    public Result returnResult() {
+        Result result = new Result();
+        result.setStatus(0);
+        result.setMessage("you haven't log in");
         return result;
     }
 }
